feat: :boom: Disable the original Improved Match Timer if we detect it running
If both plugins are running for whatever reason, this could cause issues. As we provide the same functionality if no configuration changes are present, there's no harm in doing this. It is up to providers to decide on which plugin they'd prefer to use. I hope dewbsku does not choose to be malicious and rename their plugin to circumvent this detection.